Add an Intel GPU device
The Intel GPU requires a workgroup size of maximum 256 threads.  To be
able to support this without changing the gpu hardware description, we
made sure that the translate pass works when the number of units in the
par_groups of the source hardware description exceeds that of the target
hardware description.